Move regression tests to tests folder
