+++ tests/conftest.py
@@ -1,0 +1,8 @@
+from project.conftest import (  # noqa
+    algorithm_config,
+    algorithm_network_config,
+    command_line_arguments,
+    command_line_overrides,
+    datamodule_config,
+    experiment_dictconfig,
+)
